Moving MapVis, updating ref
